Сделана первая работающая версия редактирования

+++ src/main/java/ru/fess38/finance/controller/AccountController.java
@@ -7,10 +7,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
-
 import ru.fess38.finance.dao.AccountDao;
 import ru.fess38.finance.dao.TransactionDao;
 import ru.fess38.finance.model.Account;
+import ru.fess38.finance.model.AccountType;
 
 import java.util.List;
 
@@ -49,7 +49,8 @@ public class AccountController {
 
   @RequestMapping(value = "/account/delete", method = RequestMethod.POST)
   public void delete(@RequestBody Account account) {
-    if (transactionDao.countByAccount(account) == 0) {
+    if (transactionDao.countByAccount(account) == 0
+        && account.getType() == AccountType.DEFAULT) {
       accountDao.delete(account);
     }
   }
